# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AppTrabajo API
  description: API REST para gestión de facturación, clientes, productos y cumplimiento fiscal (VeriFactu/AEAT)
  version: 1.0.0
  contact:
    name: API Support
    email: support@apptrabajo.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.apptrabajo.com
    description: Production server
  - url: http://localhost:3000
    description: Development server
...
security:
  - BearerAuth: []
paths:
...
  /api/analytics:
    get:
      tags: [Reports]
      summary: Analytics avanzados
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Datos de analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsData'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AnalyticsData:
      type: object
      properties:
        clientProfitability:
          type: array
        productProfitability:
          type: array
        cashFlow:
          type: array
        trends:
          type: array
        summary:
          type: object